pyproject.toml
[tool.poetry]
name = "image_slicer"
version = "0.1.0"
description = ""
authors = ["Chanwoong Kim <me@chanwoong.kim>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.6"
isort = "^5.8.0"
black = "^20.8b1"
mypy = "^0.790"
ptpython = "^3.0.7"
flake8 = "^3.8.4"
flake8-import-order = "^0.18.1"
flake8-pytest-style = "^1.3.0"
flake8-logging-format = "^0.6.0"
flake8-comprehensions = "^3.3.0"
flake8-builtins = "^1.5.3"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^20.11.1"
pep8-naming = "^0.11.1"
pytest = "^6.2.1"

[tool.isort]
profile = "google"
force_single_line = "False"

[tool.autopep8]
max_line_length = 79

[tool.black]
line-length = 79

[tool.poetry.scripts]
image-slicer = 'image_slicer.__main__:app'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
